src/requirements.txt
python-dotenv
dnspython
yoyo-migrations
requests
pygit2
psutil
